+++ nodeconductor/structure/models.py
@@ -137,6 +137,7 @@ class VATMixin(models.Model):
         charge = self.get_vat_charge()
         if charge.action == pyvat.VatChargeAction.charge:
             return charge.rate
+
         # Return None, if reverse_charge or no_charge action is applied
 
     def get_vat_charge(self):
@@ -840,9 +841,6 @@ class VirtualMachineMixin(CoordinatesMixin):
     min_ram = models.PositiveIntegerField(default=0, help_text='Minimum memory size in MiB')
     min_disk = models.PositiveIntegerField(default=0, help_text='Minimum disk size in MiB')
 
-    external_ips = models.GenericIPAddressField(null=True, blank=True, protocol='IPv4')
-    internal_ips = models.GenericIPAddressField(null=True, blank=True, protocol='IPv4')
-
     image_name = models.CharField(max_length=150, blank=True)
 
     key_name = models.CharField(max_length=50, blank=True)
@@ -870,6 +868,28 @@ class VirtualMachineMixin(CoordinatesMixin):
     @lru_cache(maxsize=1)
     def get_all_models(cls):
         return [model for model in apps.get_models() if issubclass(model, cls)]
+
+    @property
+    def external_ips(self):
+        """
+        Returns a list of external IPs.
+        Implementation of this method in all derived classes guarantees all virtual machine have the same interface.
+        For instance:
+         - SaltStack (aws) handles IPs as private and public IP addresses;
+         - DigitalOcean has only 1 external ip called ip_address.
+        """
+        return []
+
+    @property
+    def internal_ips(self):
+        """
+        Returns a list of internal IPs.
+        Implementation of this method in all derived classes guarantees all virtual machine have the same interface.
+        For instance:
+         - SaltStack (aws) handles IPs as private and public IP addresses;
+         - DigitalOcean does not support internal IP at the moment.
+        """
+        return []
 
 
 class PublishableMixin(models.Model):